% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{generate_colors}
\alias{generate_colors}
\title{Generate vector of colors}
\usage{
generate_colors(
  n,
  palette_function = NULL,
  colors = NULL,
  reverse_palette = FALSE
)
}
\arguments{
\item{n}{Number of colors}

\item{palette_function}{Function to generate colors (e.g. "parula")}

\item{colors}{Vector of colors (must be length >= n)}

\item{reverse_palette}{If TRUE, reverse colors}
}
\description{
Generate vector of colors
}
