% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_plot_widget.R
\name{insert_plot_widgets}
\alias{insert_plot_widgets}
\title{Insert plotWidget modules}
\usage{
insert_plot_widgets(
  data = shiny::reactive(NULL),
  cfg,
  id,
  session = shiny::getDefaultReactiveDomain(),
  global_settings = reactive(NULL),
  renderTable_args = list(digits = 1, align = "l"),
  ...
)
}
\arguments{
\item{data}{A reactive dataframe, used for all plots}

\item{cfg}{Config list (maybe read from YAML)}

\item{id}{Placeholder id where to place plotWidgets (used by insertUI)}

\item{session}{Shiny session object, no need to set.}

\item{global_settings}{Optional reactive list with settings; these override those in cfg,
and can be used to set settings that apply to all plots in a single insert.}

\item{renderTable_args}{List with components 'digits' and 'align' (not more for now), sent to renderTable}

\item{\dots}{Further arguments passed to \code{plotWidgetUI}}
}
\description{
Insert plotWidget modules
}
