% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_custom_plots.R
\docType{data}
\name{internal_custom_plot_types}
\alias{internal_custom_plot_types}
\title{Available custom plot functions in shintoviz}
\format{
An object of class \code{character} of length 3.
}
\usage{
internal_custom_plot_types
}
\description{
Available custom plot functions in shintoviz
}
\keyword{datasets}
