% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_plot_functions.R
\name{plot_bubble}
\alias{plot_bubble}
\title{Bubble plot : factor vs. factor}
\usage{
plot_bubble(
  data,
  xvar = "group1",
  yvar = "n",
  groupvar = "group2",
  xlab = xvar,
  ylab = yvar,
  title = "",
  subtitle = "",
  title_adjust = c("plot", "figure"),
  palette_function = NULL,
  colors = NULL,
  base_size = 15,
  reverse_order_group = FALSE,
  reverse_order_x = FALSE,
  ...
)
}
\arguments{
\item{data}{Dataframe}

\item{xvar}{(factor) Variable along the X-axis}

\item{yvar}{Numeric variable for size /colour of bubbled}

\item{groupvar}{(factor) variable along the Y-axis}

\item{xlab}{X-axis label}

\item{ylab}{Y-axis label}

\item{title}{Plot title}

\item{subtitle}{Plot subtitle}

\item{title_adjust}{Adjust the title to the plot or figure region (i.e. more to the left)}

\item{palette_function}{Name of 'pals' palette to fill the bubbles}

\item{colors}{Optional vector of colours}

\item{base_size}{Base text size}

\item{reverse_order_group}{Whether to reverse the ordering of the Y-axis groups}

\item{reverse_order_x}{Whether to reverse the ordering of the X-axis groups}

\item{...}{Further arguments ignored}
}
\description{
Bubble plot : factor vs. factor
}
