% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_plot_functions.R
\name{plot_distribution_by_group}
\alias{plot_distribution_by_group}
\title{TODO combine some features with plot_horizontal_bars

#' Make a standard horizontal barplot with group fill
#' @param data A dataframe made with \code{table_n_woningen}
#' @param palette_function A function that takes a single integer argument to return a vector of colors
plot_grouped_horizontal_bars <- function(data,
group1 = "group",  # group to make bars
group2 = "group2", # group within bars
yvar = "n",
palette_function,
reverse_palette = FALSE,
colors = NULL,
base_size = 14,
label_size = 4,
label_k = FALSE,
bar_width = NULL,
ylab = "ylab",
title = "",
sort = TRUE,
top_n = NA,
y_multiplier = 1.2,...){}
\usage{
plot_distribution_by_group(
  data,
  xvar = "",
  group = "",
  base_size = 15,
  ylab = "",
  xlab = "",
  n_bins = 20
)
}
\description{
data$group1 <- data[\link{group1}]
data$group2 <- data[\link{group2}]
data$n <- data[\link{yvar}]
}
\section{TODO fill with na_value ?}{
data <- filter(data, !is.na(group1))
}

\section{TODO te specifiek}{
data$n_total <- ave(data$n, data$group1, FUN = sum)

if(sort){
data <- mutate(data, group1 = reorder(group1, n_total, max))
} else {
data <- mutate(data, group1 = as.factor(group1))
}

data_total <- group_by(data, group1) \%>\%
summarize(n = sum(n), .groups = "drop") \%>\%
mutate(group1 = factor(group1, levels = levels(data$group1)))

if(!is.na(top_n)){
top_n <- as.numeric(top_n)
n <- nlevels(data$group1)
top_n <- pmin(top_n,n)

\if{html}{\out{<div class="sourceCode">}}\preformatted{i_l <- (n-top_n+1):n
toplevs <- levels(data$group1)[i_l]
data <- filter(data, group1 \%in\% toplevs)
data_total <- filter(data_total, group1 \%in\% toplevs)
}\if{html}{\out{</div>}}

}

if(nrow(data) == 0 || sum(data$n) == 0){
return(NULL)
}

g <- levels(data[\link{group2}])
colors <- generate_colors(length(g), palette_function, colors, reverse_palette)
}

\section{Om genoeg ruimte te maken voor de bar labels, dit moet algemener (maar hoe...)}{

}

