% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_plot_functions.R
\name{plot_grouped_horizontal_barplot}
\alias{plot_grouped_horizontal_barplot}
\title{Make a standard horizontal barplot with group fill}
\usage{
plot_grouped_horizontal_barplot(
  data,
  xvar = "group",
  yvar = "n",
  fillvar = xvar,
  group_format_function = NULL,
  sort = TRUE,
  top_n = NA,
  reverse_order = FALSE,
  palette_function = NULL,
  reverse_palette = FALSE,
  colors = NULL,
  base_size = 15,
  label_function = NULL,
  label_size = 5,
  label_k = FALSE,
  label_perc = FALSE,
  label_hjust = -0.2,
  bar_width = 0.6,
  title = "",
  subtitle = "",
  title_adjust = c("plot", "figure"),
  legend_text_size = 12,
  ...
)
}
\description{
Make a standard horizontal barplot with group fill
}
