% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_plot_functions.R
\name{plot_grouped_horizontal_bars}
\alias{plot_grouped_horizontal_bars}
\title{Make a standard horizontal barplot with group fill}
\usage{
plot_grouped_horizontal_bars(
  data,
  group1 = "group",
  group2 = "group2",
  yvar = "n",
  palette_function,
  reverse_palette = FALSE,
  colors = NULL,
  base_size = 14,
  label_size = 4,
  label_k = FALSE,
  bar_width = NULL,
  ylab = "ylab",
  title = "",
  sort = TRUE,
  top_n = NA,
  y_multiplier = 1.2
)
}
\arguments{
\item{data}{A dataframe made with \code{table_n_woningen}}

\item{palette_function}{A function that takes a single integer argument to return a vector of colors}
}
\description{
Make a standard horizontal barplot with group fill
}
