% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_plot_functions.R
\name{plot_grouped_value_by_time}
\alias{plot_grouped_value_by_time}
\title{Make a standard grouped time plot (lines or stacked bars)}
\usage{
plot_grouped_value_by_time(
  data,
  xvar = "time",
  yvar = "n",
  group = "group",
  plot_type = c("bars", "lines"),
  palette_function,
  colors = NULL,
  base_size = 14,
  label_size = 4,
  point_size = 3,
  line_width = 1.2,
  label_bars = FALSE,
  ylab = "ylab",
  xlab = "xlab",
  grouplab = "",
  title = "title",
  fill_na_group = "Onbekend",
  ...
)
}
\arguments{
\item{data}{A dataframe}

\item{xvar}{Name of variable in data for X-axis ("time")}

\item{yvar}{Name of variable for Y-axis}

\item{group}{Name of variable in data for groups}

\item{plot_type}{Either 'bars' or 'lines'}

\item{palette_function}{A function that takes a single integer argument to return a vector of colors}

\item{colors}{If palette_function not provided, a vector of colors (must be at least nrow(data))}

\item{base_size}{Font base size (title, axes)}

\item{label_size}{Size of text labels next to bars}

\item{ylab}{Y-axis label}

\item{xlab}{X-axis label}

\item{grouplab}{group label (for legend title)}

\item{title}{Title above plot}

\item{fill_na_group}{Value to use for missing group values in data ("Onbekend")}

\item{\dots}{Further arguments passed to `generate_colors`}

\item{label_k}{If TRUE, labels are divided by 1000}

\item{label_perc}{If TRUE, % is shown in label}

\item{bar_width}{Relative width of bars (1 = fills space completely)}
}
\description{
Make a standard grouped time plot (lines or stacked bars)
}
