% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_plot_functions.R
\name{plot_horizontal_bars}
\alias{plot_horizontal_bars}
\title{Make a standard horizontal barplot}
\usage{
plot_horizontal_bars(
  data,
  xvar = "group",
  yvar = "n",
  reverse_order = FALSE,
  palette_function = NULL,
  colors = NULL,
  base_size = 14,
  label_size = 4,
  label_k = FALSE,
  label_perc = FALSE,
  bar_width = NULL,
  title = "",
  ...
)
}
\arguments{
\item{data}{A dataframe}

\item{xvar}{Name of variable in data to determine bars}

\item{yvar}{Name of variable in data for length of bars}

\item{palette_function}{A function that takes a single integer argument to return a vector of colors}

\item{colors}{If palette_function not provided, a vector of colors (must be at least nrow(data))}

\item{base_size}{Font base size (title, axes)}

\item{label_size}{Size of text labels next to bars}

\item{label_k}{If TRUE, labels are divided by 1000}

\item{label_perc}{If TRUE, % is shown in label}

\item{bar_width}{Relative width of bars (1 = fills space completely)}

\item{title}{Title above plot}

\item{\dots}{Further arguments passed to `generate_colors`}
}
\description{
Make a standard horizontal barplot
}
