% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_plot_functions.R
\name{plot_pie_chart}
\alias{plot_pie_chart}
\title{Make a standard piechart}
\usage{
plot_pie_chart(
  data,
  xvar = "group",
  yvar = "n",
  palette_function,
  colors = NULL,
  base_size = 14,
  legend.position = "left",
  label_function = NULL,
  label_size = 4,
  label_hjust = 1.65,
  title = "title",
  ...
)
}
\arguments{
\item{data}{A dataframe}

\item{xvar}{Name of variable in data for X-axis ("time")}

\item{yvar}{Name of variable for Y-axis}

\item{palette_function}{A function that takes a single integer argument to return a vector of colors}

\item{colors}{If palette_function not provided, a vector of colors (must be at least nrow(data))}

\item{base_size}{Font base size (title, axes)}

\item{label_size}{Size of text labels next to bars}

\item{title}{Title above plot}

\item{\dots}{Further arguments passed to \code{generate_colors}}

\item{group}{Name of variable in data for groups}

\item{sub_type}{Either 'bars' or 'lines'}

\item{point_size}{Size of points plotted on line}

\item{line_width}{Width of line (if sub_type = "lines")}

\item{label_bars}{TRUE/FALSE, label the bars or not.}

\item{label_k}{If TRUE, labels are divided by 1000}

\item{label_perc}{If TRUE, \% is shown in label}

\item{bar_width}{Relative width of bars (1 = fills space completely)}

\item{xlab}{X-axis label}

\item{ylab}{Y-axis label}

\item{grouplab}{group label (for legend title)}
}
\description{
Make a standard piechart
}
