% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWidget.R
\name{plotWidgetUI}
\alias{plotWidgetUI}
\alias{plotWidgetModule}
\title{Shiny UI function for plotWidget}
\usage{
plotWidgetUI(id, header_ui = NULL, footer_ui = NULL, ...)

plotWidgetModule(
  input,
  output,
  session,
  plot_data = reactive(NULL),
  plot_type = reactive("plot_horizontal_bars"),
  settings = reactive(list()),
  table_format = function(x) x,
  extra_ggplot = reactive(NULL),
  y_min = NULL
)
}
\arguments{
\item{id}{Shiny input ID}

\item{header_ui}{Further UI to be placed above the plot}

\item{footer_ui}{Further UI to be placed below the plot}

\item{\dots}{Further arguments to softui::tab_box}
}
\description{
Shiny UI function for plotWidget
}
