% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_plot_data.R
\name{prepare_plot_data}
\alias{prepare_plot_data}
\title{Prepare a table for plotting with various options}
\usage{
prepare_plot_data(
  data,
  groupvar = "group",
  sort = FALSE,
  reverse = FALSE,
  fill_na_group = "Onbekend",
  top_n = NULL
)
}
\arguments{
\item{data}{A dataframe}
}
\description{
Further adjust a table prepared for plotting: change the order of
plotting, sort the levels (by their value), show only the first n values, etc.
}
