% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{set_plotwidget_font}
\alias{set_plotwidget_font}
\title{Set a Google font for plotWidget}
\usage{
set_plotwidget_font(googlefont)
}
\arguments{
\item{googlefont}{Name of the \href{https://fonts.google.com/}{Google font}}
}
\description{
Call this once (typically in global.R)
}
\examples{
set_plotwidget_font("Open Sans")
}
