% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_plot_functions.R
\name{plot_horizontal_bars}
\alias{plot_horizontal_bars}
\title{Make a standard horizontal barplot}
\usage{
plot_horizontal_bars(
  data,
  xvar = "group",
  yvar = "n",
  reverse_order = FALSE,
  palette_function = NULL,
  colors = NULL,
  base_size = 15,
  label_size = 5,
  label_k = FALSE,
  label_perc = FALSE,
  label_hjust = -0.2,
  bar_width = 0.6,
  title = "",
  ...
)
}
\arguments{
\item{data}{A dataframe}

\item{xvar}{Name of variable in data to determine bars}

\item{yvar}{Name of variable in data for length of bars}

\item{reverse_order}{Plot order of bars opposite to factor levels}

\item{palette_function}{A function that takes a single integer argument to return a vector of colors}

\item{colors}{If palette_function not provided, a vector of colors (must be at least nrow(data))}

\item{base_size}{Font base size (title, axes)}

\item{label_size}{Size of text labels next to bars}

\item{label_k}{If TRUE, labels are divided by 1000}

\item{label_perc}{If TRUE, percentage is shown in label}

\item{label_hjust}{Space between bar label and bar (negative value = outside bar as the default)}

\item{bar_width}{Relative width of bars (1 = fills space completely)}

\item{title}{Title above plot}

\item{\dots}{Further arguments passed to \code{generate_colors}}
}
\description{
Make a standard horizontal barplot
}
\examples{
library(gapminder)
library(dplyr)

plot_data <- gapminder \%>\%
  filter(year == 2007) \%>\%
  group_by(continent) \%>\%
  summarize(population = floor(1e-06 * sum(pop)), .groups = "drop")

plot_horizontal_bars(plot_data,
                     xvar = "continent",
                     yvar = "population",
                     palette_function = "ocean.phase",
                     colors = NULL,
                     base_size = 14,
                     label_size = 4,
                     label_k = FALSE,
                     bar_width = 0.62,
                     title = "Populatie (miljoenen)")
}
