% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_plot_functions.R
\name{plot_value_by_time}
\alias{plot_value_by_time}
\title{Make a standard line plot with time on the X axis}
\usage{
plot_value_by_time(
  data,
  xvar = "time",
  yvar = "n",
  plot_type = c("bars", "lines"),
  palette_function,
  colors = NULL,
  base_size = 14,
  label_size = 4,
  point_size = 3,
  line_width = 1.2,
  label_bars = FALSE,
  label_k = FALSE,
  label_perc = FALSE,
  ylab = "ylab",
  xlab = "xlab",
  title = "title",
  ...
)
}
\arguments{
\item{data}{A dataframe}

\item{xvar}{Name of variable in data to determine bars}

\item{yvar}{Name of variable in data for length of bars}

\item{plot_type}{Either 'bars' or 'lines'}

\item{palette_function}{A function that takes a single integer argument to return a vector of colors}

\item{colors}{If palette_function not provided, a vector of colors (must be at least nrow(data))}

\item{base_size}{Font base size (title, axes)}

\item{label_size}{Size of text labels next to bars}

\item{label_k}{If TRUE, labels are divided by 1000}

\item{label_perc}{If TRUE, % is shown in label}

\item{title}{Title above plot}

\item{\dots}{Further arguments passed to `generate_colors`}

\item{bar_width}{Relative width of bars (1 = fills space completely)}
}
\description{
No group (see plot_grouped_time_plot)
}
