% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWidget.R
\name{plotWidgetUI}
\alias{plotWidgetUI}
\alias{plotWidgetModule}
\title{Shiny UI function for plotWidget}
\usage{
plotWidgetUI(id, header_ui = NULL, footer_ui = NULL, ...)

plotWidgetModule(
  input,
  output,
  session,
  data = reactive(NULL),
  plot_type = reactive("plot_horizontal_bars"),
  settings = reactive(list()),
  extra_ggplot = reactive(NULL),
  y_min = NULL
)
}
\arguments{
\item{id}{Shiny input ID}

\item{header_ui}{Further UI to be placed above the plot}

\item{footer_ui}{Further UI to be placed below the plot}

\item{\dots}{Further arguments to softui::tab_box}

\item{input}{Shiny input (do not set)}

\item{output}{Shiny output (do not set)}

\item{session}{Shiny session (do not set)}

\item{plot_type}{Reactive plotting function. Can be one of \code{\link[=internal_custom_plot_types]{internal_custom_plot_types()}}}

\item{settings}{Reactive list of parameters passed to the plotting function (and table function)}

\item{extra_ggplot}{A reactive (can be a list) of expressions to add to the ggplot object}

\item{plot_data}{Reactive dataset used in plotting}
}
\description{
Shiny UI function for plotWidget
}
