% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWidget.R
\name{plotWidgetUI}
\alias{plotWidgetUI}
\alias{plotWidgetModule}
\title{Shiny UI function for plotWidget}
\usage{
plotWidgetUI(
  id,
  header_ui = NULL,
  footer_ui = NULL,
  ui_container = c("tab_box", "tabset_panel"),
  tab_settings = FALSE,
  ...
)

plotWidgetModule(
  input,
  output,
  session,
  data = reactive(NULL),
  plot_type = reactive("plot_horizontal_bars"),
  settings = reactive(list()),
  extra_ggplot = reactive(NULL),
  y_min = NULL
)
}
\arguments{
\item{id}{Shiny input ID}

\item{header_ui}{Further UI to be placed above the plot}

\item{footer_ui}{Further UI to be placed below the plot}

\item{tab_settings}{If TRUE, adds a settings tab (unused at the moment)}

\item{\dots}{Further arguments to \code{\link[softui:tabbox]{softui::tab_box()}}}

\item{input}{Shiny input (do not set)}

\item{output}{Shiny output (do not set)}

\item{session}{Shiny session (do not set)}

\item{data}{Reactive dataset used in plotting}

\item{plot_type}{Reactive plotting function. Can be one of \code{\link[=internal_custom_plot_types]{internal_custom_plot_types()}}}

\item{settings}{Reactive list of parameters passed to the plotting function (and table function)}

\item{extra_ggplot}{A reactive (can be a list) of expressions to add to the ggplot object}

\item{y_min}{Y-axis minimum value (often 0). TODO include more axis options}
}
\description{
Shiny UI function for plotWidget
}
\examples{
library(softui)
library(gapminder) # example data

# Set Google font
set_plotwidget_font("Roboto")


ui <- softui::simple_page(
  plotWidgetUI("plot1", width = 4)
)

server <- function(input, output, session) {


  raw_data <- reactive({
    filter(gapminder, year == 2007) \%>\%
      mutate(pop = floor(pop * 1e-06))
  })

  callModule(plotWidgetModule, "plot1",
             data = raw_data,
             plot_type = reactive("plot_horizontal_bars"),
             settings = reactive(
               list(
                 table_prepare = list(
                   fun = "prepare_grouped_data",
                   groupvar = "continent",
                   groupfun = sum,
                   yvar = "pop"),
                 xvar = "continent",
                 yvar = "pop",
                 palette_function = "parula",
                 title = "World population"
               )
             )
  )


}

#shinyApp(ui, server)
}
