% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWidget.R
\name{plotWidgetUI}
\alias{plotWidgetUI}
\alias{plotWidgetModule}
\title{Shiny UI function for plotWidget}
\usage{
plotWidgetUI(
  id,
  header_ui = NULL,
  footer_ui = NULL,
  ui_container = c("tab_box", "tabset_panel"),
  plotOutput_only = FALSE,
  height = 400,
  interactive = NULL,
  open_in_modal_button = TRUE,
  ...
)

plotWidgetModule(
  input,
  output,
  session,
  data = reactive(NULL),
  plot_type = reactive("plot_horizontal_bars"),
  settings = reactive(list()),
  global_settings = reactive(NULL),
  interactive = NULL,
  extra_ggplot = reactive(NULL),
  renderTable_args = list(digits = 1, align = "l"),
  y_min = NULL
)
}
\arguments{
\item{id}{Shiny input ID}

\item{header_ui}{Further UI to be placed above the plot}

\item{footer_ui}{Further UI to be placed below the plot}

\item{ui_container}{Either \code{tab_box} or \code{tabset_panel} to contain the plot widget}

\item{plotOutput_only}{Only make a \code{plotOutput}, not the entire widget with tabs}

\item{height}{Height of the \code{plotOutput} and \code{tableOutput}}

\item{interactive}{List with interactive controls}

\item{open_in_modal_button}{If TRUE, adds a button}

\item{\dots}{Further arguments to \code{\link[softui:tabbox]{softui::tab_box()}}}

\item{input}{Shiny input (do not set)}

\item{output}{Shiny output (do not set)}

\item{session}{Shiny session (do not set)}

\item{data}{Reactive dataset used in plotting}

\item{plot_type}{Reactive plotting function. Can be one of \code{\link[=internal_custom_plot_types]{internal_custom_plot_types()}}}

\item{settings}{Reactive list of parameters passed to the plotting function (and table function)}

\item{global_settings}{Optional reactive list with settings; these override those in settings().
Used in \code{insert_plot_widgets} to pass settings that apply to all plots at the same time.}

\item{extra_ggplot}{A reactive (can be a list) of expressions to add to the ggplot object}

\item{renderTable_args}{List of arguments (only digits and align) to send to renderTable}

\item{y_min}{Obsolete}
}
\description{
Shiny UI function for plotWidget
}
\examples{
library(softui)
library(gapminder) # example data

# Set Google font
set_plotwidget_font("Roboto")


ui <- softui::simple_page(
  plotWidgetUI("plot1", width = 4)
)

server <- function(input, output, session) {


  raw_data <- reactive({
    filter(gapminder, year == 2007) \%>\%
      mutate(pop = floor(pop * 1e-06))
  })

  callModule(plotWidgetModule, "plot1",
             data = raw_data,
             plot_type = reactive("plot_horizontal_bars"),
             settings = reactive(
               list(
                 table_prepare = list(
                   fun = "prepare_grouped_data",
                   groupvar = "continent",
                   groupfun = sum,
                   yvar = "pop"),
                 xvar = "continent",
                 yvar = "pop",
                 palette_function = "parula",
                 title = "World population"
               )
             )
  )


}

#shinyApp(ui, server)
}
